From uint32_t to double for prob function

+++ source/events/probability_space/ProbabilitySpace.cpp
@@ -1,6 +1,6 @@
 #include "source/events/probability_space/ProbabilitySpace.h"
 
-ProbabilitySpace::ProbabilitySpace(const Omega* omega, const SigmaAlgebra* sigmaAlgebra, const Function<Event, uint32_t>* probabilityFunction) {
+ProbabilitySpace::ProbabilitySpace(const Omega* omega, const SigmaAlgebra* sigmaAlgebra, const Function<Event, double>* probabilityFunction) {
     this->setOmega(omega);
     this->setProbabilityFunction(probabilityFunction);
     this->setSigmaAlgebra(sigmaAlgebra);
@@ -16,7 +16,7 @@ void ProbabilitySpace::setSigmaAlgebra(const SigmaAlgebra* sigmaAlgebra) {
     this->sigmaAlgebra = sigmaAlgebra;
 }
 
-void ProbabilitySpace::setProbabilityFunction(const Function<Event, uint32_t>* probabilityFunction) {
+void ProbabilitySpace::setProbabilityFunction(const Function<Event, double>* probabilityFunction) {
     if (!probabilityFunction) throw std::runtime_error("Nullptr detected");
     this->probabilityFunction = probabilityFunction;
 }
@@ -29,6 +29,6 @@ const SigmaAlgebra* ProbabilitySpace::getSigmaAlgebra() const {
     return this->sigmaAlgebra;
 }
 
-const Function<Event, uint32_t>* ProbabilitySpace::getProbabilityFunction() const {
+const Function<Event, double>* ProbabilitySpace::getProbabilityFunction() const {
     return this->probabilityFunction;
 }
